auth-api/requirements.txt
Flask==0.12
PyJWT==1.4.2
